Use `Foo.bar`, not `Foo::bar`, for class methods
There were ~14 instances of `ClassName::class_method_name` in the _src,
but ~602 instances of `ClassName.class_method_name`.
Though using double colons is syntatically valid, it is less used in the
community compared to just using a period.

This PR just switches all the instances I found to use periods,
for consistency throughout the site.